rename(generator): generator 패키지 util 패키지로 이동
