+++ docs/README.md
@@ -190,7 +190,7 @@
     - [x] 맞힌 숫자 개수, 당첨 금액을 가지고 있다
     - [x] 맞힌 숫자 개수를 키로 대응되는 Enum을 조회할 수 있다
 
-11. 당첨 통계 그리고 각 통계별 개수를 담는, 출력을 위한 데이터를 형성한다 - LottoService 
+11. 당첨 통계 그리고 각 통계별 개수를 담는, 출력을 위한 데이터를 형성한다 - LottoFinalResult 
   - 구입한 금액에 따라 형성된 로또 리스트와 사용자의 LottoWithBonus를 비교한다
     - [x] 로또 리스트를 형성한다 
     - [x] 생성한 로또 리스트를 Repository에 저장해야 한다
