+++ src/test/java/lotto/repository/LottoRepositoryTest.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import java.util.stream.Stream;
 import lotto.domain.Lotto;
 import lotto.util.LottoGenerator;
-import lotto.generator.MockedRandomGenerator;
+import lotto.util.generator.MockedRandomGenerator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.provider.Arguments;
